Fix bug in YGSQLite class, when add new object to database fail.

+++ Ledger/YGTools.h
@@ -41,7 +41,7 @@
 
 + (NSString *)humanViewStringForByteSize:(NSUInteger)size;
 
-+ (NSInteger)defaultFontSize;
++ (CGFloat)defaultFontSize;
 
 + (NSInteger)deviceScreenWidth;
 + (NSInteger)lengthCharachtersForTableView;
